added my stuff
also added lines to make my function run. You have the game set up for 3 players, I have it set up for 5 so we will need to decide on that. changed Player.sick to 0 so it works with my code. Changed player1, player2, and player3 to lowercase because it was easier and safer than changing all of my instances of players to caps. Also need to remember that my code asks if spare parts are in the inventory, but it is not properly referenced because spare parts are not yet coded - we need to figure out how to do that